refactor(integration): split the test into different modules
