+++ tests/integration/test_charm.py
@@ -4,73 +4,35 @@
 
 import asyncio
 import logging
-from pathlib import Path
 
 import pytest
-import yaml
-from httpx import HTTPStatusError
-from juju.model import Model
-from lightkube import Client
+from helpers import (
+    APP_NAME,
+    MISSING_RELATION_MESSAGE,
+    TIMEOUT,
+    UNTRUST_ERROR_MESSAGE,
+    get_model,
+)
 from lightkube.core.exceptions import ApiError
 from lightkube.resources.apiextensions_v1 import CustomResourceDefinition
-from lightkube.resources.apps_v1 import DaemonSet
+from lightkube.resources.apps_v1 import DaemonSet, Deployment
+from lightkube.resources.core_v1 import Secret, ServiceAccount
+from lightkube.resources.rbac_authorization_v1 import ClusterRoleBinding
 from pytest_operator.plugin import OpsTest
 
-from constants import VELERO_NODE_AGENT_NAME
-from velero import Velero
+DEPLOYMENT_IMAGE_ERROR_MESSAGE_1 = "Velero Deployment is not ready: ImagePullBackOff"
+DEPLOYMENT_IMAGE_ERROR_MESSAGE_2 = "Velero Deployment is not ready: ErrImagePull"
+USE_NODE_AGENT_CONFIG_KEY = "use-node-agent"
+VELERO_IMAGE_CONFIG_KEY = "velero-image"
+VELERO_NODE_AGENT_NAME = "node-agent"
 
 logger = logging.getLogger(__name__)
 
-TIMEOUT = 60 * 10
-USE_NODE_AGENT_CONFIG_KEY = "use-node-agent"
-VELERO_AWS_PLUGIN_IMAGE_KEY = "velero-aws-plugin-image"
-VELERO_IMAGE_CONFIG_KEY = "velero-image"
-METADATA = yaml.safe_load(Path("./charmcraft.yaml").read_text())
-APP_NAME = METADATA["name"]
-DEFAULT_VELERO_IMAGE = METADATA["config"]["options"][VELERO_IMAGE_CONFIG_KEY]["default"]
-
-S3_INTEGRATOR = "s3-integrator"
-S3_INTEGRATOR_CHANNEL = "latest/stable"
-
-UNTRUST_ERROR_MESSAGE = (
-    "The charm must be deployed with '--trust' flag enabled, run 'juju trust ...'"
-)
-READY_MESSAGE = "Unit is Ready"
-MISSING_RELATION_MESSAGE = "Missing relation: [s3-credentials]"
-DEPLOYMENT_IS_NOT_READY_MESSAGE = "Velero Deployment is not ready: "
-
-
-@pytest.fixture(scope="session")
-def lightkube_client() -> Client:
-    """Return a lightkube client to use in this session."""
-    client = Client(field_manager=APP_NAME)
-    return client
-
-
-def get_velero(model: str) -> Velero:
-    """Return a Velero instance for the given model."""
-    return Velero("./velero", model)
-
-
-def get_model(ops_test: OpsTest) -> Model:
-    """Return the Juju model of the current test.
-
-    Returns:
-        A juju.model.Model instance of the current model.
-
-    Raises:
-        AssertionError if the test doesn't have a Juju model.
-    """
-    model = ops_test.model
-    if model is None:
-        raise AssertionError("ops_test has a None model.")
-    return model
-
 
 @pytest.mark.abort_on_fail
-async def test_build_and_deploy(ops_test: OpsTest, s3_connection_info):
-    """Build the velero-operator and deploy it with the integrator charms."""
-    logger.info("Building and deploying velero-operator charm with s3-integrator")
+async def test_build_and_deploy(ops_test: OpsTest):
+    """Build and deploy the velero-operator."""
+    logger.info("Building and deploying velero-operator charm")
 
     charm = await ops_test.build_charm(".")
     model = get_model(ops_test)
@@ -79,37 +41,11 @@ async def test_build_and_deploy(ops_test: OpsTest, s3_connection_info):
         model.deploy(
             charm, application_name=APP_NAME, trust=False, config={"use-node-agent": True}
         ),
-        model.deploy(S3_INTEGRATOR, channel=S3_INTEGRATOR_CHANNEL),
-        model.wait_for_idle(apps=[APP_NAME, S3_INTEGRATOR], status="blocked", timeout=TIMEOUT),
+        model.wait_for_idle(apps=[APP_NAME], status="blocked", timeout=TIMEOUT),
     )
 
     for unit in model.applications[APP_NAME].units:
         assert unit.workload_status_message == UNTRUST_ERROR_MESSAGE
-
-
-@pytest.mark.abort_on_fail
-async def test_configure_s3_integrator(
-    ops_test: OpsTest,
-    s3_cloud_credentials,
-    s3_cloud_configs,
-):
-    """Configure the integrator charm with the credentials and configs."""
-    model = get_model(ops_test)
-
-    logger.info("Setting credentials for %s", S3_INTEGRATOR)
-    await model.applications[S3_INTEGRATOR].set_config(s3_cloud_configs)
-    action = await model.units[f"{S3_INTEGRATOR}/0"].run_action(
-        "sync-s3-credentials", **s3_cloud_credentials
-    )
-    result = await action.wait()
-    assert result.results.get("return-code") == 0
-
-    logger.info("Waiting for %s to be active", S3_INTEGRATOR)
-    await model.wait_for_idle(
-        apps=[S3_INTEGRATOR],
-        status="active",
-        timeout=TIMEOUT,
-    )
 
 
 @pytest.mark.abort_on_fail
@@ -140,6 +76,7 @@ async def test_config_use_node_agent(ops_test: OpsTest, lightkube_client):
     model = get_model(ops_test)
     app = model.applications[APP_NAME]
 
+    logger.info("Setting use-node-agent to false")
     await asyncio.gather(
         app.set_config({USE_NODE_AGENT_CONFIG_KEY: "false"}),
         model.wait_for_idle(apps=[APP_NAME], timeout=TIMEOUT),
@@ -152,6 +89,7 @@ async def test_config_use_node_agent(ops_test: OpsTest, lightkube_client):
         if ae.response.status_code != 404:
             raise ae
 
+    logger.info("Setting use-node-agent to true")
     await asyncio.gather(
         app.set_config({USE_NODE_AGENT_CONFIG_KEY: "true"}),
         model.wait_for_idle(apps=[APP_NAME], timeout=TIMEOUT),
@@ -174,16 +112,18 @@ async def test_config_velero_image(ops_test: OpsTest):
     app = model.applications[APP_NAME]
     new_image = "velero-test"
 
+    logger.info("Setting velero-image to %s", new_image)
     await app.set_config({VELERO_IMAGE_CONFIG_KEY: new_image})
     async with ops_test.fast_forward(fast_interval="60s"):
         await model.wait_for_idle(apps=[APP_NAME], timeout=TIMEOUT, status="blocked")
 
     for unit in model.applications[APP_NAME].units:
-        assert DEPLOYMENT_IS_NOT_READY_MESSAGE in unit.workload_status_message and (
-            "ImagePullBackOff" in unit.workload_status_message
-            or "ErrImagePull" in unit.workload_status_message
+        assert (
+            unit.workload_status_message == DEPLOYMENT_IMAGE_ERROR_MESSAGE_1
+            or unit.workload_status_message == DEPLOYMENT_IMAGE_ERROR_MESSAGE_2
         )
 
+    logger.info("Resetting velero-image config to default")
     await app.reset_config([VELERO_IMAGE_CONFIG_KEY])
     async with ops_test.fast_forward(fast_interval="60s"):
         await model.wait_for_idle(apps=[APP_NAME], timeout=TIMEOUT, status="blocked")
@@ -193,56 +133,12 @@ async def test_config_velero_image(ops_test: OpsTest):
 
 
 @pytest.mark.abort_on_fail
-@pytest.mark.parametrize(
-    "integrator,plugin_image_key",
-    [(S3_INTEGRATOR, VELERO_AWS_PLUGIN_IMAGE_KEY)],
-)
-async def test_integrator_relation(ops_test: OpsTest, integrator: str, plugin_image_key: str):
-    """Test the relation between the velero-operator charm and the integrator charm."""
-    model = get_model(ops_test)
-    app = model.applications[APP_NAME]
-    new_plugin_image = "velero-test-plugin-image"
-
-    logger.info("Relating velero-operator to %s", integrator)
-    await model.integrate(APP_NAME, integrator)
-    async with ops_test.fast_forward(fast_interval="60s"):
-        await model.wait_for_idle(
-            apps=[APP_NAME],
-            status="active",
-            timeout=TIMEOUT,
-        )
-    for unit in model.applications[APP_NAME].units:
-        assert unit.workload_status_message == READY_MESSAGE
-
-    await app.set_config({plugin_image_key: new_plugin_image})
-    async with ops_test.fast_forward(fast_interval="60s"):
-        await model.wait_for_idle(apps=[APP_NAME], timeout=TIMEOUT, status="blocked")
-
-    await app.reset_config([plugin_image_key])
-    async with ops_test.fast_forward(fast_interval="60s"):
-        await model.wait_for_idle(apps=[APP_NAME], timeout=TIMEOUT, status="active")
-
-    logger.info("Unrelating velero-operator from %s", integrator)
-    await ops_test.juju(*["remove-relation", APP_NAME, integrator])
-    async with ops_test.fast_forward(fast_interval="60s"):
-        await model.wait_for_idle(
-            apps=[APP_NAME],
-            status="blocked",
-            raise_on_blocked=False,
-            timeout=TIMEOUT,
-        )
-    for unit in model.applications[APP_NAME].units:
-        assert unit.workload_status_message == MISSING_RELATION_MESSAGE
-
-
-@pytest.mark.abort_on_fail
 async def test_remove(ops_test: OpsTest, lightkube_client):
     """Remove the applications and assert that all resources are deleted."""
     model = get_model(ops_test)
-    velero = get_velero(model.name)
 
+    logger.info("Removing velero-operator charm")
     await asyncio.gather(
-        model.remove_application(S3_INTEGRATOR),
         model.remove_application(APP_NAME),
         model.block_until(
             lambda: model.applications[APP_NAME].status == "unknown",
@@ -250,16 +146,16 @@ async def test_remove(ops_test: OpsTest, lightkube_client):
         ),
     )
 
-    for resource in velero._core_resources + velero._storage_provider_resources:
-        try:
-            lightkube_client.get(resource.type, resource.name)
-            assert False, f"Resource {resource.name} was not deleted"
-        except (ApiError, HTTPStatusError) as ae:
-            assert ae.response.status_code == 404
-
-    result = list(
-        lightkube_client.list(
-            CustomResourceDefinition, labels={"component": "velero"}, namespace=model.name
+    logger.info("Checking that all resources are deleted")
+    for resource in [
+        Deployment,
+        DaemonSet,
+        ServiceAccount,
+        ClusterRoleBinding,
+        Secret,
+        CustomResourceDefinition,
+    ]:
+        res = list(
+            lightkube_client.list(resource, labels={"component": "velero"}, namespace=model.name)
         )
-    )
-    assert not result, "CustomResourceDefinitions were not deleted"
+        assert not res, "Velero {} still exists".format(resource.__name__)
